Add a gridcomponent.py module
gridcomponent.py contains GridComponent class which is a base class for
components that will be placed on a grid. The class handles loading and
scaling of the component's image. It also keeps track where the
component is located on the grid, even if it is moved via one of it's
move-methods. It's also possible to transform() the component on to a new grid.