Change Placement and Clash to comply new requirements
A GridComponent class was implemented and the Ship class was changed to
inherit from the GridComponent class. The following changes were made to
comply the new requirements:

- Add the new constructor parameters for the initialization of Ship
  objects
- Change awaiting_ship to not call update_squares() in Placement's
  constructor because it's not needed anymore as the update is done
  internally in the GridComponent class
- Change the parameters of the calls to the movement methods of the Ship class
- Remove the call to the update_squares in check_events() method of the
  Placement class because it is done internally in the GridComponent
  class

Hitmarkers and Missmarkers were also added. They are used in the Clash
class. When striking a target, hitcollision is checked. If the strike
hits an enemy ship, a hitmarker will be drawn on the targeted square.
If it's a miss, a missmarker will be drawn instead.

- Add import of Hitmarker class from hitmarker.py
- Add import of Missmarker class from missmarker.py
- Add check_hitcollision() method to Clash class for checking if the
  strike hit or missed the target
- Add call to check_hitcollision() method when pressing Enter in the
  Clash scene
  - a Hitmarker object or a Missmarker object is created depending on if
  the strike hit and the created object is added to it's respective group