Add ending conditions and a base for game ending scene
A base for class End was added. This scene is for drawing the results
after the game.

Functionality was added for checking if winning conditions are met. If
all ships have been destroyed, a packet will be sent to the opponent
notifying that the game has come to an end. The sender will switch to
End scene right after sending the packet. The receiver will switch
scenes when the game over packet is received.