Change conditions of switching scenes and cut unused code
Ship positions are not sent to opponent anymore. The turn based gameplay
works by sending the results of the strike to opponent, so there is no
need to check the hits locally.

When the player has placed his/her ships, a packet is sent to the
opponent to signal that he/she is ready. The player waits for a ready
message from the opponent. After both are ready, the scene is switched.