Add length attribute
The length attribute defines the packet's intended length
defined when the packet is created or it's  data is set.
The header length is included in the attribute.

When sending the packet through the network, it is possible
that only part of the data is received. In that case, more
data should be received until the total size is the same as
the length attribute.